move `migrate-deprecated-utilities` to core
